% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{utilities}
\alias{utilities}
\alias{helper}
\title{A helper function}
\usage{
helper()
}
\description{
\code{helper()} provides an inspirational message to help
you through the day.
}
\examples{
helper()

}
